moved the getBitbucketData and retriedAllPagesBitbucket functions to a Bitbucket object. We can make similar objects for GitHub and GitLab as well
